+++ test/dummy/config/initializers/recording_studio.rb
@@ -11,7 +11,11 @@ RecordingStudio.configure do |config|
     "RecordingStudioSupport::SupportSection",
     "RecordingStudioSupport::SupportPage",
     "RecordingStudioAttachable::Attachment",
-    "RecordingStudioPublishable::Publishable"
+    "RecordingStudioPublishable::Publishable",
+    "RecordingStudio::Access",
+    "RecordingStudioApi::ApiClient",
+    "RecordingStudioApi::ApiCredential",
+    "RecordingStudioApi::ApiAccessToken"
   ]
 
   config.require_recordable_declarations = true